backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
googlemaps==4.10.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
reportlab==4.0.7
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

